Se termino el modulo de courseInstructor y se agregó el nuevo modulo de Enrollment con las validaciones respectivas de integridad sobre student y course

+++ UniversityReact.API/Controllers/StudentController.cs
@@ -115,6 +115,11 @@ namespace UniversityReact.API.Controllers
                     return BadRequest(BAD_REQUEST_ID);
                 }
 
+                if ( await student.existsStudentInOtherTable(id))
+                {
+                    return BadRequest("El estudiante se esta utilizando por lo tanto no se puede eliminar");
+                }
+
                 if (!student.DeleteStudent(id))
                 {
                     return InternalServerError(new Exception(INTERNAL_SERVER_ERROR_MSG));
